# public/assets/api.yaml
openapi: 3.0.0
info:
  title: 'swagger easycrud-ci4'
  description: 'This is a sample server easycrud-ci4'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: /api
    description: 'OpenApi host'
paths:
...
  /donaturcategories:
    get:
      tags:
        - Donaturcategory
      summary: 'Find list Donaturcategory'
      description: 'Returns list of Donaturcategory'
      operationId: getDonaturcategory
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Donaturcategory' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Donaturcategory' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Donaturcategory' } }
                type: object
        '404':
          description: 'Donaturcategory not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Donaturcategory
      summary: 'Add a new Donaturcategory to the store'
      operationId: addDonaturcategory
      requestBody:
        $ref: '#/components/requestBodies/Donaturcategory'
      responses:
        '201':
          description: 'Created Donaturcategory'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donaturcategory'
            application/xml:
              schema:
                $ref: '#/components/schemas/Donaturcategory'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/donaturcategories/{id}':
    get:
      tags:
        - Donaturcategory
      summary: 'Find Donaturcategory by ID'
      description: 'Returns a single Donaturcategory'
      operationId: getDonaturcategoryById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Donaturcategory to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donaturcategory'
            application/xml:
              schema:
                $ref: '#/components/schemas/Donaturcategory'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Donaturcategory not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Donaturcategory
      summary: 'Update an existing Donaturcategory'
      operationId: updateDonaturcategory
      parameters:
        -
          name: id
          in: path
          description: 'Donaturcategory id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Donaturcategory'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Donaturcategory not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Donaturcategory
      summary: 'Deletes a Donaturcategory'
      operationId: deleteDonaturcategory
      parameters:
        -
          name: id
          in: path
          description: 'Donaturcategory id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /donaturs:
    get:
      tags:
        - Donatur
      summary: 'Find list Donatur'
      description: 'Returns list of Donatur'
      operationId: getDonatur
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Donatur' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Donatur' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Donatur' } }
                type: object
        '404':
          description: 'Donatur not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Donatur
      summary: 'Add a new Donatur to the store'
      operationId: addDonatur
      requestBody:
        $ref: '#/components/requestBodies/Donatur'
      responses:
        '201':
          description: 'Created Donatur'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donatur'
            application/xml:
              schema:
                $ref: '#/components/schemas/Donatur'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/donaturs/{id}':
    get:
      tags:
        - Donatur
      summary: 'Find Donatur by ID'
      description: 'Returns a single Donatur'
      operationId: getDonaturById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Donatur to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Donatur'
            application/xml:
              schema:
                $ref: '#/components/schemas/Donatur'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Donatur not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Donatur
      summary: 'Update an existing Donatur'
      operationId: updateDonatur
      parameters:
        -
          name: id
          in: path
          description: 'Donatur id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Donatur'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Donatur not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Donatur
      summary: 'Deletes a Donatur'
      operationId: deleteDonatur
      parameters:
        -
          name: id
          in: path
          description: 'Donatur id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
    Donatur:
      title: Donatur
      description: Donatur
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 11
        name:
          nullable: false
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 128
        id_kategori:
          nullable: false
          title: id_kategori
          description: id_kategori
          type: integer
          format: '-'
          maxLength: 11
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: date
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: date
        created_by:
          nullable: true
          title: created_by
          description: created_by
          type: integer
          format: '-'
          maxLength: 11
        updated_by:
          nullable: true
          title: updated_by
          description: updated_by
          type: integer
          format: '-'
          maxLength: 11
      type: object
    Donaturcategory:
      title: Donaturcategory
      description: Donaturcategory
      properties:
        id:
          nullable: false
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 11
        name:
          nullable: false
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 128
        label:
          nullable: false
          title: label
          description: label
          type: string
          format: '-'
          maxLength: 255
        path_image:
          nullable: true
          title: path_image
          description: path_image
          type: string
          format: '-'
          maxLength: 255
        description:
          nullable: false
          title: description
          description: description
          type: string
          format: '-'
        created_at:
          nullable: false
          title: created_at
          description: created_at
          type: string
          format: date
        updated_at:
          nullable: false
          title: updated_at
          description: updated_at
          type: string
          format: date
        created_by:
          nullable: true
          title: created_by
          description: created_by
          type: integer
          format: '-'
          maxLength: 11
        updated_by:
          nullable: true
          title: updated_by
          description: updated_by
          type: integer
          format: '-'
          maxLength: 11
      type: object
...
  requestBodies:
    Donatur:
      description: 'Donatur object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Donatur'
        application/xml:
          schema:
            $ref: '#/components/schemas/Donatur'
        application/json:
          schema:
            $ref: '#/components/schemas/Donatur'
    Donaturcategory:
      description: 'Donaturcategory object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Donaturcategory'
        application/xml:
          schema:
            $ref: '#/components/schemas/Donaturcategory'
        application/json:
          schema:
            $ref: '#/components/schemas/Donaturcategory'
...
  securitySchemes:
    bearer_auth:
      type: http
      description: 'Login with email and password to get the authentication token'
      name: 'Token based Based'
      in: header
      bearerFormat: JWT
      scheme: bearer
...